Revert "drop 2.6.24 support"
This reverts commit dc4399ee1ee946c20f491f16ae97613609d5a939.

Although the 2.6.25 kernel has been released, it is still quite recent.
Some fairly new distributions ship with 2.6.24 (Debian testing and
Ubuntu Hardy, to name a couple). In those ones, the kernel version will
not be updated until the next distribution release. If the driver is
only compatible with 2.6.25 we are making it impossible to develop or
use the driver on those distributions, therefore reverting the patch.