Show received signal strength information
This patch allows user space tools such as iwconfig, iwlist and
NetworkManager show the relative received signal strength. It is still
incomplete in the sense that:

* The only PHY data reported by the tested acx111 chip was the RSSI, but
the mapping from RSSI units to dBm or the maximum RSSI value (in order
to estimate a percentage) is still unknown.

* The Windows XP driver (using ndiswrapper) reports quality, signal
level and noise level. The first one in percentage, the latter two in
dBm. Whether these values are right, remains to be determined. The
GPL'ed part of the AVM FRITZ!WLAN USB Stick sources show that at least
on the TNETW1450 only the RSSI value (mapped to signal level) seems to
be supported.

* There is currently a mapping from RSSI to percentage in the
acx_signal_to_winlevel() function. However, the same function in the
FreeBSD/DragonFlyBSD driver uses different calculation constants
depending on the chip type. Furthermore, the OpenBSD driver defines
maximum RSSI levels depending on the type of radio transceiver (i.e.
Maxim, Ralink, Radia, etc.). Contacting them in order to find out how
they got to those results would probably be a good idea.