Add the acx prefix to log messages
The acx: prefix has been added to all those lines where logging was
performed and it was missing. This practice, as recommended on the
comments at acx_debug.h, allows the user to discern those messages
that come from the driver from those that come from other parts of the
kernel.