Move the main config options to the Makefile
The main configuration options (i.e. whether we are compiling for a PCI
or USB device) have now been moved to the Makefile, so that there is no
need to modify the source code when using different configuration
settings at compile time.

See the comments in the Makefile for more information.