Properly add the GPL license
The original code was released under a dual GPL/MPL license. The MPL
license was removed on commit 75fbf0f7b217673cd0b7b76a5761fa8d0683c19e.

In the current commit, the copyright notice and the copying permission
statement have been added as per
http://www.gnu.org/licenses/gpl-howto.html. Notice that also the COPYING
file, which contains a copy of the GPL2 license has been added. For
in-tree drivers, it is probably not required, since there is a copy in
the kernel tree itself. But since the driver is currently not included
in mainline kernel, we should probably distribute this file.

Another question is whether we want the MPL license back. The removal on
the above mentioned commit was carried out without prior consultation to
the acx100-devel list.