Cleanups, Changelog additions, new release numbering scheme
Changed the release number series of the release: 0.5.0. In this way,
the mac80211 version has got its own numbering scheme (0.3.x is for
non-mac80211, 0.4.x was for the now dead softmac branch).

Changelog: removed the TODO tasks to the newly-created TODO file. Added
information about the latest releases, in GNU Changelog format.

Removed the README.unknowncommands file and moved back the information
elements there to acx_ie.h.

Changed the copyright notice on all files to include all years of
releases.

Cleanup: restored some of the original comments and removed the
misleading ones.