Created an INSTALL file, Makefile modifications
An INSTALL file has bee added to the driver, with detailed instructions
about building and installing it.

The Makefile has been modified in order to make it also usable for an
in-tree build.

A short notice indicating that some of the information in the README
file has been added to it. Sonn the complete README file will be
updated.