mem: Locking changes of hw data-access functions
Locking in mem and ACXMEM_WARN_NOT_SPIN_LOCKED check
====================================================

Locking in mem is more complex as for pci, because the different data-access
functions below need to be protected against incoming interrupts.

Data-access on the mem device is always going in serveral, none-atomic steps,
involving 2 or more register writes (e.g. ACX_SLV_REG_ADDR, ACX_SLV_REG_DATA).

If an interrupt is serviced while a data-access function is ongoing, this
may give access interference with by the involved operations, since the
irq routine is also using the same data-access functions.

In case of interference, this often manifests during driver operations as
failure of device cmds and subsequent hanging of the device. It especially
appeared during sw-scans while a connection was up.

For this reason, irqs shall be off while data access functions are executed,
and for this we'll use the acx-spinlock.

In pci we don't have this problem, because all data-access functions are
atomic enough and we use dma (and the sw-scan problem is also not observed in
pci, which indicates confirmation).

Apart from this, the pure acx-sem locking is already coordinating accesses
well enough, that simple driver operation without inbetween scans work without
problems.

Different locking approaches a possible to solves this (e.g. fine vs
coarse-grained).

The chosen approach is:

1) Mem.c data-access functions contain all a check to insure, they are executed
under the acx-spinlock.
=> This is the red line that tells, if something needs coverage.

2) The scope of acx-spinlocking is local, in this case here only to mem.c.
All common.c functions are already protected by the sem.

3) In order to consolidate locking calls and also to account for the logic
of the various write_flush() calls around, locking in mem should be:

a) as coarse-grained as possible, and ...

b) ... as fine-grained as required. Basically that means, that before
functions, that sleep, unlocking needs to be done. And locking is taken up
again inside the sleeping function. Specifically the cmd-functions are used
in this path.

Once stable, the locking checks in the data-access functions could be #defined
away. Mem.c is anyway more used two smaller cpus (pxa UP e.g.), so the implied
runtime constraints by the lock won't take much effect.

Signed-off-by: Oliver Winker <oliver@oli1170.net>