all modules: Remove old locking naming convention from function names
This is the first commit of a series to remove the _i, _s, _l, _e
naming convention from the old locking scheme.

Due to the new sem-based locking the naming convention is not so much
useful anymore. And also within the old locking scheme, the naming
was often not correct or missing - so finally the added value was
anyhow reduced.

Signed-off-by: Oliver Winker <oliver@oli1170.net>