strip acxmem: and acxpci: prefixes from log() formats
This cleans up format strings missed in the acx: removal patch.  Both
mem.c and pci.c had a mix of acx: and acxmem: and acxpci: prefixes, so
this patch makes things more uniform.  If acxmem: and acxpci: are
really desired throughout these files, then redeffing pr_fmt() after
including acx_debug.h will do it.

It also reduces the diffs between mem.c and pci.c (by a trivial
amount, but every bit helps), for the de-duplication patch Oliver is
aiming for.

Oliver, if you've started the de-duplication work and this gets in the
way, just drop it, its trivial to redo:

     perl -pi -e 's/, "acxmem:\s+/, "/' mem.c
     perl -pi -e 's/, "acxpci:\s+/, "/' pci.c

Signed-off-by: Jim Cromie <jim.cromie@gmail.com>
Reviewed-by: Florian Fainelli <florian@openwrt.org>