mem: refactor INLINE_IOs into mem-inlines.h
Move low-level read/write functions (all of which are static inline)
out of mem.c to a separate header file, along with the "Locking in
mem" commentary above them, and replace with an include.

Ifdef out their fn-sigs at the top of mem.c, which are somewhat
redundant.  This required moving the include line up, above
acxmem_log_txbuffer(), which uses one of the functions.  Result is
better conformance with linux bottom-up coding convention.

Note: these fns have same sig as those used by pci.c, but as the
"Locking in mem" commentary suggests, are quite different in
implementation.  Should review those diffs, add a try-both-and-compare
shim, and maybe reduce or eliminate the differences.

Signed-off-by: Jim Cromie <jim.cromie@gmail.com>