pci.c: use PCI_VDEVICE macro
Use macro to define pci-device-id table, similar to the use of
USB_DEVICE() in usb.c.  Also add #error for !CONFIG_PCI.  I tested
this with CONFIG_ACX_MAC80211_PCI=n and it works (ie pci.c is not
compiled), so this suggests that all the #if CONFIG_PCIs could be
removed.

Also adds #if0d info-table as a breadcrumb - its fairly common in
other wifi drivers, used for firmware loading.

Signed-off-by: Jim Cromie <jim.cromie@gmail.com>