acx_struct_dev.h: fold 3 adjacent ifdefs together
there were 3 adjacent pairs of ifdefs; ACX_MAC80211_PCI and
ACX_MAC80211_MEM, each defining one field.  Merge them.

This changes the layout of the struct in minor ways; one test resulted
in an OOPs, and I thought this was exposing a field position brittleness,
but I cannot replicate it after a rebuild and reboot.

Signed-off-by: Jim Cromie <jim.cromie@gmail.com>