merge: acx(pci|mem)_tx_data()
update merge.c copy of acxmem_tx_data(), and rename it to
_acx_tx_data(), leading underscore distinguishes it from the generic
wrapper in common.c, which now calls it.

Besides the usual IS_PCI|MEM branches, _acx_tx_data() has a goto
is_pci_branch to skip large block of mem-specific code.  Some of it
may be relevant to pci functionality, but it works as currently done.

other bits:
- #if0d fn-defns in pci.c, mem.c, and //= fn-protos
- rewrapped comment blocks
- added acxmem_allocate_acx_txbuf_space() fn-proto to mem.h for use in
  _acx_tx_data(), and dropped STATick

Tested-by: jimc - ifups on pci (no mem card tests, as for all previous too)

Signed-off-by: Jim Cromie <jim.cromie@gmail.com>