mem.c/mem-inlines.h: Convert mem data-access functions to work with void *iobase
This converts the mem (read|write)_reg(8|16|32)() macros and
acxmem_chaincopy_to_slavemem(), acxmem_chaincopy_from_slavemem to work
with an void *iobase, and not depend on an 'u32 iobase[]' statement anymore.

This is in preparation to drop the different membase and iobase definitions
in struct acx_device for mem and pci.

Tested on hx4700.

Signed-off-by: Oliver Winker <oliver@oli1170.net>