merge.c: acx(mem|pci)_write_fw() compile test
Merge acx(mem|pci)_write_fw() together and rename, but dont use yet.
This is simple compile-test of an ugly if-def mess.  The presence of
NOPE and FW_NO_AUTO_INCREMENT suggest uncertainty in the use of the
feature, so cautious step-by-step changes seem warranted.

Im entertaining converting these ifdefs to module parameters/options,
which would look cleaner, and thus be easier to read, and would allow
testing of both variations with same module.  But 1st, finish the
merge, as is.  Your view on this ??

Signed-off-by: Jim Cromie <jim.cromie@gmail.com>