common.c: add ifdefs to disable field-refs for USB-only compiles
The aim here is to guard adev->tx, adev->rx refs, which are not
present in acx_device_t when only USB is configured.  This is
insufficient, and there may be a better way (like ifdef'g the entire
function, or perhaps file), so do it separately so its easy to revert
later.

Signed-off-by: Jim Cromie <jim.cromie@gmail.com>