acx_compat: cleanup ifdefs in op_add|remove_interface
In acx_compat.h, add a kernel-version dependent macro (2.6.34) which
maps ieee80211_vif --> ieee80211_if_init_conf.

This is a rather bold rename, and hasnt been tested on an old kernel,
but the old name on longer exists in 3.4-rc, so its a reasonable bet.

The rename allows to hide version diffs affecting the fn-prototypes
(in subject), and buries the var-name diff, which allows deletion of
several #if/#else/#end chunks in the fn-bodies.  One is left, cuz it
has a different field-name in the 2 version.

Signed-off-by: Jim Cromie <jim.cromie@gmail.com>