merge.h: make USB-only cleanly
With this patch, USB-only compiles cleanly.  In effect, it adds stub
functions (static inlines) for every function that compiler complained
about.

It does so via a new macro: decl_or_stub(proto, stubret), which may be
too-clever-by-half (or not clever enough).  The macro expands either
as a prototype, or with a "static inline" prefix and an "empty" body.

The main shortcoming with the new macro is that it must be redefined
for each conditional combination desired (only 1 is needed here),
because macros cannot contain #if constructs.  There may be a high-fu
way to do this better, please suggest it if you know how.  Despite its
weakness, it works, and eliminates a large repetetive block of
fn-protos, so keep it.

Signed-off-by: Jim Cromie <jim.cromie@gmail.com>