add DECL_OR_STUB macro to acx.h, use it in merge.h
Create a better version of decl_or_stub macro added previously to
merge.h.  This version accepts an arg which chooses whether result
expands to a declaration (of a function defined elsewhere), or a
static inline stub function.  Its purpose is to avoid double entries
in 2 legs of a big #if/#else/#endif.

Signed-off-by: Jim Cromie <jim.cromie@gmail.com>