pci.h: drop acxpci_upload_fw() prototypes
This function was moved to merge.c, so its prototype is determined by
both MEM and PCI.  This makes it fit better in merge.h than pci.h.  As
it turns out, we dont need a prototype at all (its only called from
within merge.c, but the current prototype is wrong in a USB-only
config, so just drop it.

Signed-off-by: Jim Cromie <jim.cromie@gmail.com>