merge.c: Use common acx_wait_cmd_status() for pci and mem
Use a udelay() for mem- and msleep() for the pci-case.

In AP-mode we need more often execute fw-command while up, e.g. for TIM
updates. Since AP is mostly relevant for pci and in pci we also can sleep
while do fw-commands, let use the sleep. This way the cpu can still be
scheduled to other tasks in the meanwhile.

Signed-off-by: Oliver Winker <oliver@oli1170.net>